pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thermocalc"
version = "0.1.0"
description = "Repartition approximative de consommation de chauffage a partir de tetes thermostatiques Zigbee TRV26"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116,<1.0",
  "paho-mqtt>=2.1,<3.0",
  "itsdangerous>=2.2,<3.0",
  "uvicorn[standard]>=0.35,<1.0",
  "jinja2>=3.1,<4.0",
  "pydantic>=2.11,<3.0",
  "python-multipart>=0.0.20,<1.0",
  "reportlab>=4.4,<5.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28,<1.0",
  "pytest>=8.3,<9.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
